@@ -16,7 +16,7 @@
 
 import 'package:archive/archive.dart';
 
-import '../document.dart';
+import '../format/object_base.dart';
 
 /// Zip compression function
 DeflateCallback defaultDeflate = const ZLibEncoder().encode;
