@@ -14,17 +14,20 @@
  * limitations under the License.
  */
 
-import '../data_types.dart';
 import '../document.dart';
+import '../format/array.dart';
+import '../format/bool.dart';
+import '../format/dict.dart';
+import '../format/name.dart';
 import '../graphic_state.dart';
 import 'font.dart';
-import 'object_dict.dart';
+import 'object.dart';
 import 'pattern.dart';
 import 'shading.dart';
 import 'xobject.dart';
 
 /// Helper functions for graphic objects
-mixin PdfGraphicStream on PdfObjectDict {
+mixin PdfGraphicStream on PdfObject<PdfDict> {
   /// Isolated transparency: If this flag is true, objects within the group
   /// shall be composited against a fully transparent initial backdrop;
   /// if false, they shall be composited against the group’s backdrop
@@ -48,6 +51,10 @@ mixin PdfGraphicStream on PdfObjectDict {
 
   /// The xobjects or other images in the pdf
   final xObjects = <String, PdfXObject>{};
+
+  bool _altered = false;
+  bool get altered => _altered;
+  set altered(bool _) => _altered = true;
 
   /// Add a font to this graphic object
   void addFont(PdfFont font) {
@@ -98,12 +105,14 @@ mixin PdfGraphicStream on PdfObjectDict {
     // This holds any resources for this page
     final resources = PdfDict();
 
-    resources['/ProcSet'] = PdfArray(const <PdfName>[
-      PdfName('/PDF'),
-      PdfName('/Text'),
-      PdfName('/ImageB'),
-      PdfName('/ImageC'),
-    ]);
+    if (altered) {
+      resources['/ProcSet'] = PdfArray(const <PdfName>[
+        PdfName('/PDF'),
+        PdfName('/Text'),
+        PdfName('/ImageB'),
+        PdfName('/ImageC'),
+      ]);
+    }
 
     // fonts
     if (fonts.isNotEmpty) {
@@ -127,7 +136,7 @@ mixin PdfGraphicStream on PdfObjectDict {
 
     if (pdfDocument.hasGraphicStates) {
       // Declare Transparency Group settings
-      params['/Group'] = PdfDict({
+      params['/Group'] = PdfDict.values({
         '/Type': const PdfName('/Group'),
         '/S': const PdfName('/Transparency'),
         '/CS': const PdfName('/DeviceRGB'),
@@ -138,15 +147,17 @@ mixin PdfGraphicStream on PdfObjectDict {
       resources['/ExtGState'] = pdfDocument.graphicStates.ref();
     }
 
-    if (params.containsKey('/Resources')) {
-      final res = params['/Resources'];
-      if (res is PdfDict) {
-        res.merge(resources);
-        return;
+    if (resources.isNotEmpty) {
+      if (params.containsKey('/Resources')) {
+        final res = params['/Resources'];
+        if (res is PdfDict) {
+          res.merge(resources);
+          return;
+        }
       }
-    }
 
-    params['/Resources'] = resources;
+      params['/Resources'] = resources;
+    }
   }
 }
 
