@@ -15,10 +15,14 @@
  */
 
 import '../color.dart';
-import '../data_types.dart';
 import '../document.dart';
+import '../format/array.dart';
+import '../format/dict.dart';
+import '../format/name.dart';
+import '../format/num.dart';
+import '../format/string.dart';
 import '../rect.dart';
-import 'object_dict.dart';
+import 'object.dart';
 import 'page.dart';
 
 /// Outline mode
@@ -46,7 +50,7 @@ enum PdfOutlineStyle {
 }
 
 /// Pdf Outline object
-class PdfOutline extends PdfObjectDict {
+class PdfOutline extends PdfObject<PdfDict> {
   /// Constructs a Pdf Outline object.
   /// When selected, the specified region is displayed.
   PdfOutline(
@@ -61,7 +65,7 @@ class PdfOutline extends PdfObjectDict {
     PdfPage? page,
   })  : assert(anchor == null || (dest == null && rect == null)),
         _page = page,
-        super(pdfDocument);
+        super(pdfDocument, params: PdfDict());
 
   /// This holds any outlines below us
   List<PdfOutline> outlines = <PdfOutline>[];
@@ -122,10 +126,10 @@ class PdfOutline extends PdfObjectDict {
 
     // These are for kids only
     if (parent != null) {
-      params['/Title'] = PdfSecString.fromString(this, title!);
+      params['/Title'] = PdfString.fromString(title!);
 
       if (color != null) {
-        params['/C'] = PdfColorType(color!);
+        params['/C'] = PdfArray.fromColor(color!);
       }
 
       if (style != PdfOutlineStyle.normal) {
@@ -133,7 +137,7 @@ class PdfOutline extends PdfObjectDict {
       }
 
       if (anchor != null) {
-        params['/Dest'] = PdfSecString.fromString(this, anchor!);
+        params['/Dest'] = PdfString.fromString(anchor!);
       } else {
         final dests = PdfArray();
         dests.add(dest!.ref());
