@@ -14,13 +14,17 @@
  * limitations under the License.
  */
 
-import '../data_types.dart';
 import '../document.dart';
+import '../format/array.dart';
+import '../format/dict.dart';
+import '../format/indirect.dart';
+import '../format/name.dart';
+import '../format/num.dart';
 import '../graphics.dart';
 import '../page_format.dart';
 import 'annotation.dart';
 import 'graphic_stream.dart';
-import 'object_dict.dart';
+import 'object.dart';
 import 'object_stream.dart';
 
 /// Page rotation
@@ -39,7 +43,7 @@ enum PdfPageRotation {
 }
 
 /// Page object, which will hold any contents for this page.
-class PdfPage extends PdfObjectDict with PdfGraphicStream {
+class PdfPage extends PdfObject<PdfDict> with PdfGraphicStream {
   /// This constructs a Page object, which will hold any contents for this
   /// page.
   PdfPage(
@@ -49,7 +53,12 @@ class PdfPage extends PdfObjectDict with PdfGraphicStream {
     int? index,
     int? objser,
     int objgen = 0,
-  }) : super(pdfDocument, type: '/Page', objser: objser, objgen: objgen) {
+  }) : super(pdfDocument,
+            params: PdfDict.values({
+              '/Type': const PdfName('/Page'),
+            }),
+            objser: objser,
+            objgen: objgen) {
     if (index != null) {
       pdfDocument.pdfPageList.pages.insert(index, this);
     } else {
@@ -64,10 +73,12 @@ class PdfPage extends PdfObjectDict with PdfGraphicStream {
   PdfPageRotation rotate;
 
   /// This holds the contents of the page.
-  final contents = <PdfObjectStream>[];
+  final contents = <PdfObject>[];
 
   /// This holds any Annotations contained within this page.
   final annotations = <PdfAnnot>[];
+
+  final _contentGraphics = <PdfObject, PdfGraphics>{};
 
   /// This returns a [PdfGraphics] object, which can then be used to render
   /// on to this page. If a previous [PdfGraphics] object was used, this object
@@ -76,6 +87,7 @@ class PdfPage extends PdfObjectDict with PdfGraphicStream {
   PdfGraphics getGraphics() {
     final stream = PdfObjectStream(pdfDocument);
     final g = PdfGraphics(this, stream.buf);
+    _contentGraphics[stream] = g;
     contents.add(stream);
     return g;
   }
@@ -100,8 +112,15 @@ class PdfPage extends PdfObjectDict with PdfGraphicStream {
     params['/MediaBox'] =
         PdfArray.fromNum(<double>[0, 0, pageFormat.width, pageFormat.height]);
 
+    for (final content in contents) {
+      if (!_contentGraphics[content]!.altered) {
+        content.inUse = false;
+      }
+    }
+
     // The graphic operations to draw the page
-    final contentList = PdfArray.fromObjects(contents);
+    final contentList =
+        PdfArray.fromObjects(contents.where((e) => e.inUse).toList());
 
     if (params.containsKey('/Contents')) {
       final prevContent = params['/Contents']!;
@@ -117,16 +136,17 @@ class PdfPage extends PdfObjectDict with PdfGraphicStream {
 
     if (contentList.values.length == 1) {
       params['/Contents'] = contentList.values.first;
-    } else if (contents.isNotEmpty) {
+    } else if (contentList.isNotEmpty) {
       params['/Contents'] = contentList;
     }
 
     // The /Annots object
     if (annotations.isNotEmpty) {
       if (params.containsKey('/Annots')) {
-        final annotsList = params['/Annots'];
-        if (annotsList is PdfArray) {
-          annotsList.values.addAll(PdfArray.fromObjects(annotations).values);
+        final annotationList = params['/Annots'];
+        if (annotationList is PdfArray) {
+          annotationList.values
+              .addAll(PdfArray.fromObjects(annotations).values);
         }
       } else {
         params['/Annots'] = PdfArray.fromObjects(annotations);
